Small grammatical change + consistency/clarity fix
ln 2: changed the "Find a bug with a "needsdiagnosis" label or that is otherwise new" to be the same as the corresponding step in "How to Reproduce the Bug" to make that instruction easier to understand, and be consistent in terms of linking to the "Need Diagnosis" tag rather than sending users to a label. 

ln 6;8-9: added the between "if bug" to both be consistent with the other instructions that refer to the bug as "the bug" rather than just "bug" and for grammatical correctness. 

ln 11: It's more correct to say "if there are any errors" than "if there are errors"